renamed model class files
